refactor: rename api file to service
