+++ src/hooks/index.ts
@@ -1,3 +1,5 @@
-import { useGetUser, useGetUsers } from './useUserApi';
+import { useGetUser, useGetUsers } from './useUserService';
+import { useGetGroups } from './useGroupService';
 
 export { useGetUser, useGetUsers };
+export { useGetGroups };
